# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Stratum API
  description: |
    A code collaboration platform for the AI engineering era.

    This specification is generated from the route code (`src/index.ts` and
    `src/routes/*.ts`) and covers every `/api/*` route, the root health check,
    and the GitHub webhook receiver.

    Not covered here:
    - HTML page routes (`/`, `/p/*`, `/changes/*`, `/auth/*`, `/dev-login`, `/ui.css`, ...).
    - The git smart-HTTP endpoints (`/@{namespace}/{slug}/info/refs`,
      `git-upload-pack`, `git-receive-pack`) — documented separately in
      `docs/adr/005`.

    Authentication: most endpoints accept a Stratum API token (user or agent)
    as a `Authorization: Bearer <token>` header. Browser requests may instead be
    authenticated by the `stratum_session` cookie. Read endpoints on projects
    with `visibility: public` also accept anonymous requests. Admin endpoints
    additionally accept an `X-Admin-API-Key` header.

    Several form-friendly endpoints return a `302` redirect instead of JSON when
    the request body is form-encoded rather than `application/json`; this spec
    documents the JSON behaviour.
  version: 1.0.0
servers:
  - url: https://your-instance.workers.dev
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /api/projects/{namespace}/{slug}/issues:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Slug"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Open an issue
      description: Anyone who can read the project can open issues. Also accepts form bodies (which redirect on success).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 200 }
                body: { type: string, maxLength: 20000 }
                linkedChangeId:
                  type: string
                  description: Must reference a change in this project
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                type: object
                properties:
                  issue: { $ref: "#/components/schemas/Issue" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/TargetDeleting"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listIssues
      tags: [issues]
      summary: List issues
      description: |
        Filterable by status, label, assignee, and free-text search (`q`, a
        case-insensitive substring match over title + body). Paginate with
        `limit` + `offset`; results are ordered newest-first by issue number.
      security: [{ bearerAuth: [] }, {}]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, closed] }
        - name: label
          in: query
          description: Only issues carrying this exact label
          schema: { type: string }
        - name: assignee
          in: query
          description: Only issues assigned to this user id
          schema: { type: string }
        - name: q
          in: query
          description: Substring search over title and body (max 200 chars; % and _ match literally)
          schema: { type: string, maxLength: 200 }
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 500 }
        - name: offset
          in: query
          description: Rows to skip (LIMIT/OFFSET pagination)
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items: { $ref: "#/components/schemas/Issue" }
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{namespace}/{slug}/issues/{number}:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Slug"
      - name: number
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    get:
      operationId: getIssue
      tags: [issues]
      summary: Issue detail
      security: [{ bearerAuth: [] }, {}]
      responses:
        "200":
          description: Issue
          content:
            application/json:
              schema:
                type: object
                properties:
                  issue: { $ref: "#/components/schemas/Issue" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Edit, close, reopen, label, or assign an issue
      description: Requires project write access. Users only (agent tokens cannot edit issues).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
                body: { type: string, maxLength: 20000 }
                status: { type: string, enum: [open, closed] }
                linkedChangeId:
                  type: [string, "null"]
                  description: Set to null or "" to unlink
                assignee:
                  type: [string, "null"]
                  description: Single assignee (user id); set to null or "" to unassign
                labels:
                  type: array
                  maxItems: 20
                  items: { type: string, maxLength: 50 }
                  description: |
                    Replaces the issue's full label set (add/remove by sending the
                    desired set); [] clears all labels. Omit to leave labels unchanged.
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                type: object
                properties:
                  issue: { $ref: "#/components/schemas/Issue" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{namespace}/{slug}/issues/{number}/comments:
    parameters:
      - $ref: "#/components/parameters/Namespace"
      - $ref: "#/components/parameters/Slug"
      - name: number
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    post:
      operationId: addIssueComment
      tags: [issues]
      summary: Comment on an issue
      description: Anyone who can read the project can comment. Also accepts form bodies (which redirect to the issue page).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 20000 }
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment: { $ref: "#/components/schemas/IssueComment" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listIssueComments
      tags: [issues]
      summary: List an issue's comments
      description: Chronological (oldest first), paginated with limit + offset.
      security: [{ bearerAuth: [] }, {}]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 500 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items: { $ref: "#/components/schemas/IssueComment" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/changes/{id}:
    parameters:
      - $ref: "#/components/parameters/ChangeId"
    get:
      operationId: getChange
      tags: [changes]
      summary: Get a change with its eval runs and cost summary
      security: [{ bearerAuth: [] }, {}]
      responses:
        "200":
          description: Change detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  change: { $ref: "#/components/schemas/Change" }
                  evalRuns:
                    type: array
                    items: { $ref: "#/components/schemas/EvalRun" }
                  costs:
                    type: array
                    items:
                      type: object
                      description: Per-kind cost summary for this change
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Stratum API token (user token or agent token)
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Admin API key for administrator endpoints
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      schema: { type: string }
      description: Project namespace, including the @ prefix (e.g. "@alice")
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
      description: Project slug
...
    ChangeId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Change id
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (also returned for access denied, to avoid disclosing existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TargetDeleting:
      description: The project (or its owner) is being deleted (code TARGET_DELETING)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    EvalRun:
      type: object
      required: [id, changeId, evaluatorType, score, passed, reason, ranAt]
      properties:
        id: { type: string }
        changeId: { type: string }
        evaluatorType:
          type: string
          description: e.g. secret_scan, diff, webhook, llm, sandbox
        score: { type: number }
        passed: { type: boolean }
        reason: { type: string }
        issues:
          type: array
          items: { type: string }
        ranAt: { type: string, format: date-time }
    Issue:
      type: object
      required: [id, project, number, title, status, authorType, authorId, createdAt, updatedAt]
      properties:
        id: { type: string }
        project: { type: string }
        projectId: { type: string, format: uuid }
        number: { type: integer }
        title: { type: string }
        body: { type: string }
        status:
          type: string
          enum: [open, closed]
        authorType:
          type: string
          enum: [user, agent]
        authorId: { type: string }
        assignee:
          type: string
          description: Single assignee (user id); absent when unassigned
        labels:
          type: array
          items: { type: string }
          description: Returned by the issue list, detail, and update endpoints
        linkedChangeId: { type: string }
        closedAt: { type: string, format: date-time }
        closedBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    IssueComment:
      type: object
      required: [id, issueId, authorType, authorId, body, createdAt]
      properties:
        id: { type: string }
        issueId: { type: string }
        authorType:
          type: string
          enum: [user, agent]
        authorId: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
...